Can now randomly populate negative training class
In the current iteration of the project, the image parser will make a negative "not meme" class that consists of many images located in one folder.  However, the parser will output these results in an order corresponding with the search. For example, If I gave the parser terms for "forest, tank, bear", the parser will give me 100 forest pictures, 100 tank pictures, and 100 bear pictures. 

This method will move those pictures to another folder in random order. This is useful if I do not want to use all of the parsers search results but want a varied randomized data set.